# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi V1 Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Demo API Server"
        },
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Development Environment"
        },
        {
            "url": "http://127.0.0.1:9000/api/",
            "description": "Staging  Environment"
        }
    ],
    "paths": {
        "/api/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Post your email and password and we will return a token. Use the token in the 'Authorization' header like so 'Bearer YOUR_TOKEN'",
                "description": "Post your email and password and we will return a token.",
                "operationId": "login",
                "requestBody": {
                    "description": "The Token Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "your@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "YOUR_PASSWORD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Rehim Alcanov",
                                                "email": "rehimalcanov@mail.ru",
                                                "updated_at": "2022-07-20T10:57:41.000000Z",
                                                "created_at": "2022-07-20T10:57:41.000000Z",
                                                "id": "4"
                                            }
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "5|J9pEuB8PtCd2gNY24kkMLRthbbfI884UE0be7QSc"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    },
                    "401": {
                        "description": "The provided credentials are unauthorized."
                    }
                }
            }
        },
        "/api/v1/auth/user": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get user by id",
                "description": "Returns a single new user.",
                "operationId": "getUserByToken",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "type": "object",
                                            "example": {
                                                "id": "4",
                                                "name": "Rehim Alcanov",
                                                "email": "rehimalcanov@mail.ru",
                                                "email_verified_at": "2022-07-20T10:24:56.000000Z",
                                                "updated_at": "2022-07-20T10:57:41.000000Z",
                                                "created_at": "2022-07-20T10:57:41.000000Z"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The provided credentials are unauthorized."
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/logout": {
            "delete": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout User",
                "description": "This can only be done by the logged in user.",
                "operationId": "logoutUser",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Post your email and password and we will return a token. Use the token in the 'Authorization' header like so 'Bearer YOUR_TOKEN'",
                "description": "Post your email and password and we will return a token.",
                "operationId": "register",
                "requestBody": {
                    "description": "The Token Request",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "your@email.com"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "your@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "YOUR_PASSWORD"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Rehim Alcanov",
                                                "email": "rehimalcanov@mail.ru",
                                                "updated_at": "2022-07-20T10:57:41.000000Z",
                                                "created_at": "2022-07-20T10:57:41.000000Z",
                                                "id": "4"
                                            }
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "5|J9pEuB8PtCd2gNY24kkMLRthbbfI884UE0be7QSc"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "The provided credentials are incorrect."
                    },
                    "401": {
                        "description": "The provided credentials are unauthorized."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}